packages/templates/src/code-review-agent/templates/langchain-python/requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Data handling and validation
pydantic>=2.0.0

# File and git operations
python-dotenv>=1.0.0
gitpython>=3.1.40
pathlib2>=2.3.7

# Code syntax highlighting and parsing
pygments>=2.17.0

# HTTP requests and web security
requests>=2.31.0
urllib3>=2.0.0

# Cryptography and security
cryptography>=41.0.0
bcrypt>=4.0.0

# Regular expressions and pattern matching
regex>=2023.0.0

# Date and time handling
python-dateutil>=2.8.0

# JSON and data processing
jsonschema>=4.17.0

# Optional but recommended: Enhanced code analysis
# Uncomment these for additional analysis capabilities
# pylint>=3.0.0              # Static code analysis
# bandit>=1.7.0              # Security linter for Python
# mypy>=1.8.0                # Type checking
# flake8>=6.0.0              # Code style checker
# black>=23.0.0              # Code formatter
# isort>=5.12.0              # Import sorter

# Optional: AST parsing and analysis
ast-monitor>=0.4.0           # AST analysis utilities

# Optional: Additional security tools
# safety>=2.3.0              # Check for known security vulnerabilities
# semgrep>=1.45.0            # Static analysis for security patterns

# Optional: Performance analysis
# memory-profiler>=0.61.0    # Memory usage profiling
# line-profiler>=4.0.0       # Line-by-line profiling

# Development and testing (for template development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and monitoring
structlog>=23.2.0

# Optional: Additional language support
# For JavaScript/TypeScript analysis:
# nodejs>=0.1.1
#
# For Java analysis:
# javalang>=0.13.0
#
# For Go analysis:
# go-parser>=0.1.0